### 2022-01-21 #4084 @whedon
[PRE REVIEW]: SGMCMCJax: a lightweight JAX library for stochastic gradient Markov chain Monte Carlo algorithms
**Submitting author:** @jeremiecoullon (<a href="http://orcid.org/0000-0002-7032-3425">Jeremie Coullon</a>)
**Repository:** <a href="https://github.com/jeremiecoullon/SGMCMCJax" target ="_blank">https://github.com/jeremiecoullon/SGMCMCJax</a>
**Version:** 0.2.9
