### 2022-01-21 #4084 @kyleniemeyer
Hi @jeremiecoullon, while we are getting started, could you take a look at that potentially missing DOI? I also noticed some formatting issues with your references, if you could take a look at those too.

For reference, the available citation commands are:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
