### 2022-01-22 a9c04d2
fixed ref formatting issue and added missing doi
