+++ paper.bib
@@ -27,22 +27,18 @@
 }
 
 @article{baker2019sgmcmc,
-title = "sgmcmc: An R Package for Stochastic Gradient Markov Chain Monte Carlo",
-abstract = "This paper introduces the R package sgmcmc; which can be used for Bayesian inference on problems with large datasets using stochastic gradient Markov chain Monte Carlo (SGMCMC). Traditional Markov chain Monte Carlo (MCMC) methods, such as Metropolis-Hastings, are known to run prohibitively slowly as the dataset size increases. SGMCMC solves this issue by only using a subset of data at each iteration. SGMCMC requires calculating gradients of the log likelihood and log priors, which can be time consuming and error prone to perform by hand. The sgmcmc package calculates these gradients itself using automatic differentiation, making the implementation of these methods much easier. To do this, the package uses the software library TensorFlow, which has a variety of statistical distributions and mathematical operations as standard, meaning a wide class of models can be built using this framework. SGMCMC has become widely adopted in the machine learning literature, but less so in the statistics community. We believe this may be partly due to lack of software; this package aims to bridge this gap.",
-keywords = "stat.CO, stat.AP, stat.ML",
-author = "Jack Baker and Paul Fearnhead and Fox, {Emily B.} and Nemeth, {Christopher John}",
-year = "2019",
-month = oct,
-day = "31",
-doi = "10.18637/jss.v091.i03",
-language = "English",
-volume = "91",
-pages = "1--27",
-journal = "Journal of Statistical Software",
-issn = "1548-7660",
-publisher = "University of California at Los Angeles",
-number = "3",
-
+title = {sgmcmc: An R Package for Stochastic Gradient Markov Chain Monte Carlo},
+author = {Jack Baker and Paul Fearnhead and Emily B. Fox and Christopher Nemeth},
+year = {2019},
+month = {oct},
+day = {31},
+doi = {10.18637/jss.v091.i03},
+volume = {91},
+pages = {1--27},
+journal = {Journal of Statistical Software},
+issn = {1548-7660},
+publisher = {University of California at Los Angeles},
+number = {3},
 }
 
 @misc{tensorflow2015-whitepaper,
