+++ paper.md
@@ -35,7 +35,7 @@ Markov Chain Monte Carlo (MCMC) is currently the gold standard for approximating
 
 # Statement of need
 
-SGMCMCJax is a Python package written in the popular JAX library [@jax2018github]. Although there are libraries for SGMCMC algorithms in other languages and automatic differentiation frameworks [@baker2019sgmcmc, @tensorflow2015-whitepaper], there is no mature library for the JAX ecosystem. However as this has become a popular framework for machine learning and scientific computing, this gap has become more noticeable. As SGMCMC algorithms are a standard tool to train Bayesian neural networks as well as statistical models with large datasets, we have written this library of samplers to fill this gap.
+SGMCMCJax is a Python package written in the popular JAX library [@jax2018github]. Although there are libraries for SGMCMC algorithms in other languages and automatic differentiation frameworks [@baker2019sgmcmc; @tensorflow2015-whitepaper], there is no mature library for the JAX ecosystem. However as this has become a popular framework for machine learning and scientific computing, this gap has become more noticeable. As SGMCMC algorithms are a standard tool to train Bayesian neural networks as well as statistical models with large datasets, we have written this library of samplers to fill this gap.
 
 
 SGMCMCJax uses JAX to perform automatic differentiation and compilation to XLA. The use of JAX allows the SGMCMCJax library to effortlessly run on GPUs and TPUs, which is essential for large models such as BNNs. As a result, the library uses an easy-to-use interface and provides very competitive speed performance. SGMCMCJax is designed in a modular framework allowing users to simply run one of its many algorithms or to create new algorithms for research purposes by using the existing algorithms as building blocks. Furthermore, SGMCMCJax can integrate easily with other codebases within the JAX ecosystem such as Flax, a neural network library for JAX.
