### 2022-01-22 #4084 @jeremiecoullon
@kyleniemeyer : thanks for these tips!
- I fixed the reference formatting issue (Baker et al, 2019, Abadi et al 2015). Is this what you meant?
- I added the missing DOI for Baker et al (2019)

Let me know if there are any other issues I can help with
