### 2022-01-30 #4113 @whedon
[REVIEW]: SGMCMCJax: a lightweight JAX library for stochastic gradient Markov chain Monte Carlo algorithms
**Submitting author:** @jeremiecoullon (<a href="http://orcid.org/0000-0002-7032-3425">Jeremie Coullon</a>)
**Repository:** https://github.com/jeremiecoullon/SGMCMCJax
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.12
