### 2022-02-10 jeremiecoullon/SGMCMCJax#26 @canyon289
Make file hides testing failures 
Hello! I'm reviewing thiis library as part of https://github.com/openjournals/joss-reviews/issues/4113

In regards to testing I have two comments
1. I forgot to run `python setup.py develop` but when running the make file I don't receive any exceptions. If I run pytest independently then I see things are failing
2. I would suggest implementing a CI pipeline using github actions, or other similar service, to run tests when code is merged on main at a minimum. and ideally on each pull request as well.
   * If i carefully pick apart the JOSS requirement "**Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?". There are no automated steps (manual steps are provided)
