### 2022-02-10 jeremiecoullon/SGMCMCJax#27 @canyon289
Statement of problem solved
From https://github.com/openjournals/joss-reviews/issues/4113
>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The README describes who the target audience is, but doesn't quite articulate the problem the software was designed to solve. I assume this is solving the problem of people who want to define their models in JAX, and not a full blown PPL, but still be able too obtain samples? A sentence, or even a partial sentence would provide much clarity
