### 2022-02-10 jeremiecoullon/SGMCMCJax#28 @canyon289
Plot labels in documentation are hard to read
Black theme with black axes label on transparent background makes it challenging to see the full figure.
![image](https://user-images.githubusercontent.com/7213793/153336440-d316e91a-9a43-475b-bdf6-dbcdd342b58e.png)

https://github.com/openjournals/joss-reviews/issues/4113
