### 2022-02-10 jeremiecoullon/SGMCMCJax#29 @canyon289
Are community guidelines present?
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I can't seem to find clear community guidelines, save this page on how to add a sampler
https://sgmcmcjax.readthedocs.io/en/latest/add_a_sampler.html

Per JOSS ask I suggest adding a CONTRIBUTING.md guide, or equivalent, that describes the "human process" of how to submit changes, report issues, or where to reach out to get more help. If you'd like I'm happy to explain what tools ArviZ and PyMC use for these purposes.

https://github.com/openjournals/joss-reviews/issues/4113
