### 2022-02-10 jeremiecoullon/SGMCMCJax#30 @canyon289
Functionality documentation, both in docs and in code
I feel this library would benefit from API documentation, detailing
* Which methods are available for public use
* What their arguments are
* What the return values are

This library would also benefit from code documentation in the codebase itself. For esample here
https://github.com/jeremiecoullon/SGMCMCJax/blob/master/sgmcmcjax/kernels.py#L13

Both suggestions could be solved in one shot by documenting the code in a standard documentation format and then using sphinx's automatic API documentation functionality.

Do note this library has excellent examples and applied documentation.

https://github.com/openjournals/joss-reviews/issues/4113
