### 2022-02-10 #4113 @canyon289
@jeremiecoullon
I left my initial set of comments. Do know that the issue tickets are biased towards issues. Overall the library is quite easy to navigate and understand given all the code examples. I also am quite excited to see a comprehensive library of SHGMCMC implementations, and the hope that others will contribute more as the field advances.

For the issues reported I see no major deficiencies, just smaller additions or fixes that are needed in a couple of spots. I'm happy add any feedback additional feedback if you need it, or to discuss further if you would find that helpful.

Remaining in my review is the assessment of functionality. I plan to get to it soon but am not anticipating any issues.
