### 2022-02-10 9e69d37
setup API docs (basic setup)
