+++ docs/util.rst
@@ -1,0 +1,5 @@
+Utility functions
+=================
+
+.. automodule:: sgmcmcjax.util
+   :members:
