+++ docs/kernels.rst
@@ -1,0 +1,6 @@
+Kernels
+=======
+
+
+.. automodule:: sgmcmcjax.kernels
+   :members:
