+++ docs/gradient_estimation.rst
@@ -1,0 +1,7 @@
+Gradient estimation 
+=================== 
+
+
+
+.. automodule:: sgmcmcjax.gradient_estimation
+   :members:
