+++ docs/ksd.rst
@@ -1,0 +1,5 @@
+Kernel Stein Discrepancy 
+======================== 
+
+.. automodule:: sgmcmcjax.ksd
+   :members:

[…]
