### 2022-02-10 89ea9c8
added a bunch of docstrings
