+++ docs/conf.py
@@ -13,6 +13,8 @@
 import os
 import sys
 
+import sphinx_rtd_theme
+
 sys.path.insert(0, os.path.abspath("../.."))
 
 
@@ -56,7 +58,8 @@ exclude_patterns = ["build", "Thumbs.db", ".DS_Store"]
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "furo"
+# html_theme = "furo"
+html_theme = "sphinx_rtd_theme"
 
 html_theme_options = {
     "light_css_variables": {
