+++ sgmcmcjax/ksd.py
@@ -1,9 +1,18 @@
-from jax import jit, vmap, lax
 import jax.numpy as jnp
+from jax import jit, lax, vmap
+
 from .types import Array
 
+
 @jit
-def k_0_fun(parm1: Array, parm2: Array, gradlogp1: Array, gradlogp2:Array , c: float = 1., beta: float = -0.5) -> float:
+def k_0_fun(
+    parm1: Array,
+    parm2: Array,
+    gradlogp1: Array,
+    gradlogp2: Array,
+    c: float = 1.0,
+    beta: float = -0.5,
+) -> float:
     """KSD kernel with the IMQ kernel and the 2 norm: http://proceedings.mlr.press/v70/gorham17a/gorham17a.pdf
 
     Args:
@@ -17,18 +26,19 @@ def k_0_fun(parm1: Array, parm2: Array, gradlogp1: Array, gradlogp2:Array , c: f
     Returns:
         float: [description]
     """
-    diff = parm1-parm2
+    diff = parm1 - parm2
     dim = parm1.shape[0]
-    base = (c**2 + jnp.dot(diff, diff))
-    term1 = jnp.dot(gradlogp1,gradlogp2)*base**beta
-    term2 = -2*beta * jnp.dot(gradlogp1, diff) * base**(beta-1)
-    term3 = 2*beta * jnp.dot(gradlogp2, diff) * base**(beta-1)
-    term4 = -2*dim*beta*(base**(beta-1))
-    term5 = -4*beta* (beta-1)*base**(beta-2)*jnp.sum(jnp.square(diff))
+    base = c ** 2 + jnp.dot(diff, diff)
+    term1 = jnp.dot(gradlogp1, gradlogp2) * base ** beta
+    term2 = -2 * beta * jnp.dot(gradlogp1, diff) * base ** (beta - 1)
+    term3 = 2 * beta * jnp.dot(gradlogp2, diff) * base ** (beta - 1)
+    term4 = -2 * dim * beta * (base ** (beta - 1))
+    term5 = -4 * beta * (beta - 1) * base ** (beta - 2) * jnp.sum(jnp.square(diff))
     return term1 + term2 + term3 + term4 + term5
 
 
-batch_k_0_fun_rows = jit(vmap(k_0_fun, in_axes=(None,0,None,0,None,None)))
+_batch_k_0_fun_rows = jit(vmap(k_0_fun, in_axes=(None, 0, None, 0, None, None)))
+
 
 @jit
 def imq_KSD(sgld_samples: Array, sgld_grads: Array) -> float:
@@ -41,13 +51,15 @@ def imq_KSD(sgld_samples: Array, sgld_grads: Array) -> float:
     Returns:
         float: [description]
     """
-    c, beta = 1., -0.5
+    c, beta = 1.0, -0.5
     N = sgld_samples.shape[0]
 
     def body_ksd(le_sum, x):
         my_sample, my_grad = x
-        le_sum += jnp.sum(batch_k_0_fun_rows(my_sample, sgld_samples, my_grad, sgld_grads, c, beta))
+        le_sum += jnp.sum(
+            _batch_k_0_fun_rows(my_sample, sgld_samples, my_grad, sgld_grads, c, beta)
+        )
         return le_sum, None
 
-    le_sum, _ = lax.scan(body_ksd, 0., (sgld_samples, sgld_grads))
-    return jnp.sqrt(le_sum)/N
+    le_sum, _ = lax.scan(body_ksd, 0.0, (sgld_samples, sgld_grads))
+    return jnp.sqrt(le_sum) / N
