+++ sgmcmcjax/gradient_estimation.py
@@ -1,17 +1,32 @@
-from typing import Any, Tuple, Callable
 from collections import namedtuple
+from typing import Any, Callable, Tuple
 
 import jax.numpy as jnp
-from jax import random, jit, lax
-from jax.tree_util import tree_flatten, tree_unflatten, tree_map, tree_multimap
-from .types import PyTree, PRNGKey, SamplerState, SVRGState
+from jax import jit, lax, random
+from jax.tree_util import tree_flatten, tree_map, tree_multimap, tree_unflatten
+
+from .types import PRNGKey, PyTree, SamplerState, SVRGState
 
 
 # standard gradient estimator
-def build_gradient_estimation_fn(grad_log_post: Callable, data: Tuple, batch_size: int) -> Tuple[Callable, Callable]:
+def build_gradient_estimation_fn(
+    grad_log_post: Callable, data: Tuple, batch_size: int
+) -> Tuple[Callable, Callable]:
+    """Build a standard gradient estimator
+
+    Args:
+        grad_log_post (Callable): [description]
+        data (Tuple): [description]
+        batch_size (int): [description]
+
+    Returns:
+        Tuple[Callable, Callable]: [description]
+    """
     assert type(data) == tuple
     N_data, *_ = data[0].shape
-    data = tuple([jnp.array(elem) for elem in data]) # this makes sure data has jax arrays rather than numpy arrays
+    data = tuple(
+        [jnp.array(elem) for elem in data]
+    )  # this makes sure data has jax arrays rather than numpy arrays
 
     def init_gradient(key: PRNGKey, param: PyTree) -> Tuple[PyTree, SVRGState]:
         idx_batch = random.choice(key=key, a=jnp.arange(N_data), shape=(batch_size,))
@@ -19,25 +34,42 @@ def build_gradient_estimation_fn(grad_log_post: Callable, data: Tuple, batch_siz
         param_grad = grad_log_post(param, *minibatch_data)
         return param_grad, SVRGState()
 
-
     @jit
-    def estimate_gradient(i: int, key: PRNGKey, param: PyTree, svrg_state: SVRGState = SVRGState()) -> Tuple[PyTree, SVRGState]:
-       if (batch_size is None) or batch_size == N_data:
-           return grad_log_post(param, *data), svrg_state
-       else:
-           return init_gradient(key, param)
+    def estimate_gradient(
+        i: int, key: PRNGKey, param: PyTree, svrg_state: SVRGState = SVRGState()
+    ) -> Tuple[PyTree, SVRGState]:
+        if (batch_size is None) or batch_size == N_data:
+            return grad_log_post(param, *data), svrg_state
+        else:
+            return init_gradient(key, param)
 
     return estimate_gradient, init_gradient
 
+
 # Control variates
-def build_gradient_estimation_fn_CV(grad_log_post: Callable, data: Tuple, batch_size: int, centering_value: PyTree) -> Tuple[Callable, Callable]:
+def build_gradient_estimation_fn_CV(
+    grad_log_post: Callable, data: Tuple, batch_size: int, centering_value: PyTree
+) -> Tuple[Callable, Callable]:
+    """Build a Control Variates gradient estimator
+
+    Args:
+        grad_log_post (Callable): [description]
+        data (Tuple): [description]
+        batch_size (int): [description]
+        centering_value (PyTree): [description]
+
+    Returns:
+        Tuple[Callable, Callable]: [description]
+    """
     assert type(data) == tuple
     N_data, *_ = data[0].shape
-    data = tuple([jnp.array(elem) for elem in data]) # this makes sure data has jax arrays rather than numpy arrays
+    data = tuple(
[…]
