### 2022-02-11 jeremiecoullon/SGMCMCJax#33 @ColCarroll
Clarify authorship
Hello! I'm reviewing this library as part of https://github.com/openjournals/joss-reviews/issues/4113

I notice @jeremiecoullon listed as "co-first author" -- I would guess this is a typo (or surprise from formatting, or I don't understand the implication!)

![image](https://user-images.githubusercontent.com/2295568/153531429-b7922661-7202-4794-8151-0420efded9e5.png)
