### 2022-02-11 jeremiecoullon/SGMCMCJax#33 @jeremiecoullon
Oops I think I just reused that from the template (which probably had more than 1 first authors). I'll change it simply to "first author" :)
