### 2022-02-13 388eaa3
added the rest of the docstrings and built the API documentation
