+++ sgmcmcjax/kernels.py
@@ -4,9 +4,12 @@ import jax.numpy as jnp
 from jax import jit, lax, random
 
 from .diffusions import badodab, baoab, psgld, sghmc, sgld, sgldAdam, sgnht
-from .gradient_estimation import (SVRGState, build_gradient_estimation_fn,
-                                  build_gradient_estimation_fn_CV,
-                                  build_gradient_estimation_fn_SVRG)
+from .gradient_estimation import (
+    SVRGState,
+    build_gradient_estimation_fn,
+    build_gradient_estimation_fn_CV,
+    build_gradient_estimation_fn_SVRG,
+)
 from .types import DiffusionState, PRNGKey, PyTree, SamplerState, SVRGState
 from .util import build_grad_log_post, run_loop
 
@@ -32,7 +35,7 @@ def _build_langevin_kernel(
         init_gradient (Callable): initialise the state of the gradient
 
     Returns:
-        Tuple[ Callable[[PRNGKey, PyTree], SamplerState], Callable[[int, PRNGKey, SamplerState], SamplerState], Callable[[SamplerState], PyTree] ]: [description]
+        Tuple[ Callable[[PRNGKey, PyTree], SamplerState], Callable[[int, PRNGKey, SamplerState], SamplerState], Callable[[SamplerState], PyTree] ]: An (init_fun, kernel, get_params) triple.
     """
 
     # Check whether the diffusion is a palindrome (ie: splitting scheme)
@@ -86,14 +89,14 @@ def _build_sghmc_kernel(
         init_fn_diffusion (Callable): create the initial state of the diffusion
         update_diffusion (Any): updates the state of the diffusion
         get_params_diffusion (Callable): gets the parameters from the state of the diffusion
-        resample_momentum (Callable): [description]
-        estimate_gradient (Callable): [description]
-        init_gradient (Callable): [description]
+        resample_momentum (Callable): function that resamples momentum
+        estimate_gradient (Callable): gradient estimation function
+        init_gradient (Callable): function to initialise the state of the gradient
         L (int): number of leapfrog steps
         compiled_leapfrog (bool, optional): whether or not the integrator is compiled. Defaults to True.
 
     Returns:
-        Tuple[ Callable[[PRNGKey, PyTree], SamplerState], Callable[[int, PRNGKey, SamplerState], SamplerState], Callable[[SamplerState], PyTree] ]: [description]
+        Tuple[ Callable[[PRNGKey, PyTree], SamplerState], Callable[[int, PRNGKey, SamplerState], SamplerState], Callable[[SamplerState], PyTree] ]: An (init_fun, kernel, get_params) triple.
     """
 
     init_fn, langevin_kernel, get_params = _build_langevin_kernel(
@@ -130,11 +133,11 @@ def build_sgld_kernel(
     """Build kernel for SGLD
 
     Args:
-        dt (float): [description]
-        loglikelihood (Callable): [description]
-        logprior (Callable): [description]
-        data (Tuple): [description]
-        batch_size (int): [description]
+        dt (float): step size
+        loglikelihood (Callable): log-likelihood for a single data point
+        logprior (Callable): log-prior for a single data point
+        data (Tuple): tuple of data. It should either have a single array (for unsupervised problems) or have two arrays (for supervised problems)
+        batch_size (int): batch size
 
     Returns:
         Tuple[ Callable, Callable, Callable ]: An (init_fun, kernel, get_params) triple.
@@ -160,12 +163,12 @@ def build_sgldCV_kernel(
     """Build SGLD-CV kernel
 
     Args:
-        dt (float): [description]
-        loglikelihood (Callable): [description]
-        logprior (Callable): [description]
[…]
