+++ sgmcmcjax/samplers.py
@@ -6,32 +6,40 @@ import jax.numpy as jnp
 from jax import jit, lax, random
 from tqdm.auto import tqdm
 
-from .kernels import (build_badodab_kernel, build_badodabCV_kernel,
-                      build_baoab_kernel, build_psgld_kernel,
-                      build_sghmc_kernel, build_sghmc_SVRG_kernel,
-                      build_sghmcCV_kernel, build_sgld_kernel,
-                      build_sgld_SVRG_kernel, build_sgldCV_kernel,
-                      build_sgnht_kernel, build_sgnhtCV_kernel)
+from .kernels import (
+    build_badodab_kernel,
+    build_badodabCV_kernel,
+    build_baoab_kernel,
+    build_psgld_kernel,
+    build_sghmc_kernel,
+    build_sghmc_SVRG_kernel,
+    build_sghmcCV_kernel,
+    build_sgld_kernel,
+    build_sgld_SVRG_kernel,
+    build_sgldCV_kernel,
+    build_sgnht_kernel,
+    build_sgnhtCV_kernel,
+)
 from .types import DiffusionState, PRNGKey, PyTree, SamplerState, SVRGState
 from .util import progress_bar_scan
 
 
 def _build_compiled_sampler(
     init_fn: Callable[[PRNGKey, PyTree], SamplerState],
-    my_kernel: Callable[[int, PRNGKey, SamplerState], SamplerState],
+    kernel: Callable[[int, PRNGKey, SamplerState], SamplerState],
     get_params: Callable[[SamplerState], PyTree],
     pbar: bool = True,
 ) -> Callable:
     """Build generic compiled sampler
 
     Args:
-        init_fn (Callable[[PRNGKey, PyTree], SamplerState]): [description]
-        my_kernel (Callable[[int, PRNGKey, SamplerState], SamplerState]): [description]
-        get_params (Callable[[SamplerState], PyTree]): [description]
-        pbar (bool, optional): [description]. Defaults to True.
+        init_fn (Callable[[PRNGKey, PyTree], SamplerState]): function to initialise the state of chain
+        kernel (Callable[[int, PRNGKey, SamplerState], SamplerState]): transition kernel
+        get_params (Callable[[SamplerState], PyTree]): functions that gets the target parameters from the state
+        pbar (bool, optional): whether or not to display the progress bar. Defaults to True.
 
     Returns:
-        Callable: [description]
+        Callable: sampling function with the same signature as kernel
     """
 
     @partial(jit, static_argnums=(1,))
@@ -39,7 +47,7 @@ def _build_compiled_sampler(
         def body(carry, i):
             key, state = carry
             key, subkey = random.split(key)
-            state = my_kernel(i, subkey, state)
+            state = kernel(i, subkey, state)
             return (key, state), get_params(state)
 
         key, subkey = random.split(key)
@@ -54,20 +62,20 @@ def _build_compiled_sampler(
 
 def _build_noncompiled_sampler(
     init_fn: Callable[[PRNGKey, PyTree], SamplerState],
-    my_kernel: Callable[[int, PRNGKey, SamplerState], SamplerState],
+    kernel: Callable[[int, PRNGKey, SamplerState], SamplerState],
     get_params: Callable[[SamplerState], PyTree],
     pbar: bool = True,
 ) -> Callable:
     """Build generic non-compiled sampler
 
     Args:
-        init_fn (Callable[[PRNGKey, PyTree], SamplerState]): [description]
-        my_kernel (Callable[[int, PRNGKey, SamplerState], SamplerState]): [description]
-        get_params (Callable[[SamplerState], PyTree]): [description]
-        pbar (bool, optional): [description]. Defaults to True.
+        init_fn (Callable[[PRNGKey, PyTree], SamplerState]): function to initialise the state of chain
[…]
