+++ docs/conf.py
@@ -25,7 +25,7 @@ copyright = "2021, Jeremie Coullon"
 author = "Jeremie Coullon"
 
 # The full version, including alpha/beta/rc tags
-release = "0.2.9"
+release = "0.2.10"
 
 
 # -- General configuration ---------------------------------------------------
