+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.9
+current_version = 0.2.10
 
 [aliases]
 test = pytest
