### 2022-02-14 jeremiecoullon/SGMCMCJax#38 @ColCarroll
API documentation missing on website
Looks like #32 you're maybe aware of this, but this might help track that, e.g., https://sgmcmcjax.readthedocs.io/en/latest/kernels.html is currently empty.
