### 2022-02-16 3998c4b
docs: added another path to see if autodocs find sgmcmcjax
