+++ docs/conf.py
@@ -16,6 +16,7 @@ import sys
 import sphinx_rtd_theme
 
 sys.path.insert(0, os.path.abspath("../.."))
+sys.path.insert(0, os.path.abspath("../../.."))
 
 
 # -- Project information -----------------------------------------------------
