+++ docs/kernels.rst
@@ -1,6 +1,5 @@
 Kernels
 =======
 
-
 .. automodule:: sgmcmcjax.kernels
    :members:
