+++ docs/gradient_estimation.rst
@@ -1,7 +1,5 @@
 Gradient estimation 
 =================== 
 
-
-
 .. automodule:: sgmcmcjax.gradient_estimation
    :members:
