+++ docs/gradient_estimation.rst
@@ -1,5 +1,7 @@
 Gradient estimation 
 =================== 
 
+
+
 .. automodule:: sgmcmcjax.gradient_estimation
    :members:
