### 2022-02-16 011cdd4
docs: added a bunch of paths
