+++ docs/conf.py
@@ -17,7 +17,9 @@ import sphinx_rtd_theme
 
 sys.path.insert(0, os.path.abspath("../.."))
 sys.path.insert(0, os.path.abspath("../"))
-
+sys.path.append(os.path.abspath('..'))
+sys.path.append(os.path.abspath('.'))
+sys.path.insert(0, os.path.abspath("../../.."))
 
 # -- Project information -----------------------------------------------------
 
