### 2022-02-16 jeremiecoullon/SGMCMCJax#38 @jeremiecoullon
This is an issue with readthedocs not finding sgmcmcjax (but it worked locally).
Now it's finding sgmcmcjax, but not jax; so it seems that it's not installing any of the requirements. I'll investigate!
