+++ .readthedocs.yaml
@@ -1,0 +1,16 @@
+# Required
+version: 2
+
+build:
+  os: ubuntu-20.04
+  tools:
+    python: "3.9"
+
+# sphinx:
+#   configuration: docs/conf.py
+#   fail_on_warning: false
+
+python:
+  install:
+    - requirements: docs/requirements.txt
+    - requirements: requirements.txt
