### 2022-02-24 6026268
clarify that the library is about gradient based methods. Clarify the bit about innovations
