+++ paper.md
@@ -31,7 +31,8 @@ bibliography: paper.bib
 
 In Bayesian inference the _posterior distribution_ is the probability distribution over the model parameters resulting from the prior distribution and the likelihood. One can compute integrals over this distribution to obtain quantities of interest, such as the posterior mean and variance, or credible uncertainty regions. However as these integrals are often intractable for problems of interest they therefore require numerical methods to approximate them.
 
-Markov Chain Monte Carlo (MCMC) is currently the gold standard for approximating integrals needed in Bayesian inference. However as these algorithms become prohibitively expensive for large datasets, stochastic gradient MCMC (SGMCMC) [@nemeth2021stochastic] is a popular approach to approximate these integrals in these cases. This class of scalable algorithms uses data subsampling techniques to approximate gradient based sampling algorithms. Innovations in these algorithms includes using novel gradient estimation techniques, designing more efficient diffusions, and finding more stable numerical discretisations to the diffusions. This results in a wide range of algorithms that are regularly used to fit statistical models or Bayesian neural networks (BNNs). SGMCMCJax is a lightweight library that implements many SGMCMC algorithms in the literature that can easily be used for both research purposes and practical applications.
+Markov Chain Monte Carlo (MCMC) is currently the gold standard for approximating integrals needed in Bayesian inference. However as these algorithms become prohibitively expensive for large datasets, stochastic gradient MCMC (SGMCMC) [@nemeth2021stochastic] is a popular approach to approximate these integrals in these cases. This class of scalable algorithms uses data subsampling techniques to approximate gradient based sampling algorithms, and are regularly used to fit statistical models or Bayesian neural networks (BNNs). The SGMCMC literature develops new algorithms by finding novel gradient estimation techniques, designing more efficient diffusions, and finding more stable numerical discretisations to these diffusions. SGMCMCJax is a lightweight library that is designed to allow the user to innovate along these lines or use one of the gradient-based SGMCMC algorithms already included in the library. This makes SGMCMCJax very well suited for both research purposes and practical applications.
+
 
 # Statement of need
 
