+++ paper.bib
@@ -18,6 +18,20 @@
   publisher={Taylor \& Francis}
 }
 
+@inproceedings{NIPS2015_Ma_complete_recipe,
+ author = {Ma, Yi-An and Chen, Tianqi and Fox, Emily},
+ booktitle = {Advances in Neural Information Processing Systems},
+ editor = {C. Cortes and N. Lawrence and D. Lee and M. Sugiyama and R. Garnett},
+ pages = {},
+ publisher = {Curran Associates, Inc.},
+ title = {A Complete Recipe for Stochastic Gradient MCMC},
+ url = {https://proceedings.neurips.cc/paper/2015/file/9a4400501febb2a95e79248486a5f6d3-Paper.pdf},
+ volume = {28},
+ year = {2015}
+}
+
+
+
 @software{jax2018github,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
