### 2022-02-24 c9be609
paper: K Murphy book reference fix
