+++ paper.bib
@@ -132,11 +132,10 @@ author={
   year = {2018},
 }
 
-@book{Murphy2012,
-author = {Murphy, Kevin P.},
-title = {Machine Learning: A Probabilistic Perspective},
-year = {2012},
-isbn = {0262018020},
-publisher = {The MIT Press},
-abstract = {Today's Web-enabled deluge of electronic data calls for automated methods of data analysis. Machine learning provides these, developing methods that can automatically detect patterns in data and then use the uncovered patterns to predict future data. This textbook offers a comprehensive and self-contained introduction to the field of machine learning, based on a unified, probabilistic approach. The coverage combines breadth and depth, offering necessary background material on such topics as probability, optimization, and linear algebra as well as discussion of recent developments in the field, including conditional random fields, L1 regularization, and deep learning. The book is written in an informal, accessible style, complete with pseudo-code for the most important algorithms. All topics are copiously illustrated with color images and worked examples drawn from such application domains as biology, text processing, computer vision, and robotics. Rather than providing a cookbook of different heuristic methods, the book stresses a principled model-based approach, often using the language of graphical models to specify models in a concise and intuitive way. Almost all the models described have been implemented in a MATLAB software package--PMTK (probabilistic modeling toolkit)--that is freely available online. The book is suitable for upper-level undergraduates with an introductory-level college math background and beginning graduate students. }
+@unpublished{murphy2023 ,
+      title={Probabilistic Machine Learning: Advanced Topics},
+      author={Kevin Murphy},
+      year={2023},
+      url = {https://probml.github.io/pml-book/book2.html},
+note = "In preparation",
 }
