### 2022-02-24 jeremiecoullon/SGMCMCJax#33 @jeremiecoullon
Fixed
