+++ .github/workflows/makefile.yml
@@ -1,0 +1,19 @@
+name: Makefile CI
+
+on: [push]
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+    
+    - name: configure
+      run: ./configure
+      
+    - name: Install dependencies
+      run: python setup.py develop; make
+      
+    - name: Run check
+      run: make 
