### 2022-02-27 48979d0
makefile.yaml: remove .configure
