+++ .github/workflows/makefile.yml
@@ -8,9 +8,6 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
-    
-    - name: configure
-      run: ./configure
       
     - name: Install dependencies
       run: python setup.py develop; make
