+++ .github/workflows/makefile.yml
@@ -10,7 +10,7 @@ jobs:
     - uses: actions/checkout@v2
       
     - name: Install dependencies
-      run: python setup.py develop; make
+      run: python setup.py develop --user; make
       
     - name: Run check
       run: make 
