### 2022-02-27 f43b528
update makefile.yaml; install dev requirements
