+++ .github/workflows/makefile.yml
@@ -10,7 +10,10 @@ jobs:
     - uses: actions/checkout@v2
       
     - name: Install dependencies
-      run: python setup.py develop --user; make
+      run: |
+        python setup.py develop --user
+        pip install -r requirements-dev.txt
+        make
       
     - name: Run check
       run: make 
