### 2022-02-27 5d20a57
makefile.yaml: add pytests
