+++ docs/conf.py
@@ -17,8 +17,8 @@ import sphinx_rtd_theme
 
 sys.path.insert(0, os.path.abspath("../.."))
 sys.path.insert(0, os.path.abspath("../"))
-sys.path.append(os.path.abspath('..'))
-sys.path.append(os.path.abspath('.'))
+sys.path.append(os.path.abspath(".."))
+sys.path.append(os.path.abspath("."))
 sys.path.insert(0, os.path.abspath("../../.."))
 
 # -- Project information -----------------------------------------------------
