+++ .github/workflows/makefile.yaml
@@ -24,4 +24,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - uses: psf/black@stable
+    
+    run: |
+      pip install -r requirements-dev.txt
+      black --diff --check $(git ls-files '*.py')
