### 2022-02-27 08d7135
makefile.yaml: remove lint job
