### 2022-02-27 97953be
makefile.yaml: re-added lint job with better formatting
