+++ .github/workflows/makefile.yaml
@@ -20,3 +20,13 @@ jobs:
         make
         make test
   
+  lint:
+    
+    runs-on: ubuntu-latest
+    
+    steps:
+    - uses: actions/checkout@v2
+    - name: code quality checks
+      run: |
+        pip install -r requirements-dev.txt
+        black --diff --check $(git ls-files '*.py')
