+++ .github/workflows/makefile.yaml
@@ -17,8 +17,7 @@ jobs:
       
     - name: Run tests
       run: |
-        make
-        make test
+        pytest -v sgmcmcjax tests
   
   lint:
     
@@ -29,4 +28,6 @@ jobs:
     - name: code quality checks
       run: |
         pip install -r requirements-dev.txt
+        mypy sgmcmcjax
         black --diff --check $(git ls-files '*.py')
+        isort .
