### 2022-02-27 d06e3f9
makefile.yaml: moved mypy to build job
