### 2022-02-27 2d3ab14
makefile.yaml: try cd into sgmcmcjax for isort command
