+++ sgmcmcjax/examples/sgmcmc_nuts_demo.py
@@ -2,8 +2,7 @@
 
 # Compare NUTS, SGLD and Adam on sampling from a multivariate Gaussian
 from collections import namedtuple
-from typing import (Any, Callable, Dict, List, NamedTuple, Optional, Tuple,
-                    Union)
+from typing import Any, Callable, Dict, List, NamedTuple, Optional, Tuple, Union
 
 import jax.numpy as jnp
 import optax
