+++ .github/workflows/makefile.yaml
@@ -26,7 +26,8 @@ jobs:
     steps:
     - uses: actions/checkout@v2
     - name: code quality checks
-      run: pip install -r requirements-dev.txt
-      run: mypy sgmcmcjax
-      run: black --diff --check $(git ls-files '*.py')
-      run: isort .
+      run: |
+        pip install -r requirements-dev.txt
+        mypy sgmcmcjax
+        black --diff --check $(git ls-files '*.py')
+        isort .
