### 2022-02-27 68b905f
makefile.yaml: mypy needs more dependencies
