### 2022-02-27 8bcdfb2
makefile.yaml: refactor
