+++ .github/workflows/makefile.yaml
@@ -3,6 +3,7 @@ name: Makefile CI
 on: [push]
 jobs:
   build:
+    name: build
 
     runs-on: ubuntu-latest
 
@@ -13,23 +14,28 @@ jobs:
       run: |
         python setup.py develop --user
         pip install -r requirements-dev.txt
-        
+  
+  tests:
+    name: run tests
+    needs: [build]
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
       
     - name: Run tests
       run: |
         pytest -v sgmcmcjax tests
-  
+
   lint:
+    name: linting
+    needs: [build]
     
     runs-on: ubuntu-latest
     
     steps:
     - uses: actions/checkout@v2
-
-    - name: Install dependencies
-      run: |
-        python setup.py develop --user
-        pip install -r requirements-dev.txt
 
     - name: run mypy
       run: mypy sgmcmcjax
