### 2022-02-27 0dffe15
makefile: runs tests and mypy by defualt
