### 2022-02-27 99ace6b
Setup CI (#44)
this addresses issue #26: added makefile.yaml to set up CI, and fixed makefile issue
