### 2022-02-27 jeremiecoullon/SGMCMCJax#26 @jeremiecoullon
Hello! I just merged some code into the `dev` branch which addresses this issue.

Let me know if you have more questions or issues about this!
