### 2022-02-27 15e3b51
new installation instructions and test coverage
