### 2022-02-27 097d2de
new installation instructions and test coverage (#45)
