### 2022-02-27 1a43e5e
Dev (#46)
* Setup CI (#44)
* new installation instructions and test coverage (#45)
