+++ sgmcmcjax/kernels.py
@@ -146,8 +146,9 @@ def build_sgld_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn(
         grad_log_post, data, batch_size
     )
+    init_diff, update_diff, get_p_diff = sgld(dt)
     init_fn, sgld_kernel, get_params = _build_langevin_kernel(
-        *sgld(dt), estimate_gradient, init_gradient
+        init_diff, update_diff, get_p_diff, estimate_gradient, init_gradient
     )
     return init_fn, sgld_kernel, get_params
 
@@ -177,8 +178,9 @@ def build_sgldCV_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn_CV(
         grad_log_post, data, batch_size, centering_value
     )
+    init_diff, update_diff, get_p_diff = sgld(dt)
     init_fn, sgldCV_kernel, get_params = _build_langevin_kernel(
-        *sgld(dt), estimate_gradient, init_gradient
+        init_diff, update_diff, get_p_diff, estimate_gradient, init_gradient
     )
     return init_fn, sgldCV_kernel, get_params
 
@@ -208,8 +210,9 @@ def build_sgld_SVRG_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn_SVRG(
         grad_log_post, data, batch_size, update_rate
     )
+    init_diff, update_diff, get_p_diff = sgld(dt)
     init_fn, sgldSVRG_kernel, get_params = _build_langevin_kernel(
-        *sgld(dt), estimate_gradient, init_gradient
+        init_diff, update_diff, get_p_diff, estimate_gradient, init_gradient
     )
     return init_fn, sgldSVRG_kernel, get_params
 
@@ -241,8 +244,9 @@ def build_psgld_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn(
         grad_log_post, data, batch_size
     )
+    init_diff, update_diff, get_p_diff = psgld(dt, alpha, eps)
     init_fn, sgld_kernel, get_params = _build_langevin_kernel(
-        *psgld(dt, alpha, eps), estimate_gradient, init_gradient
+        init_diff, update_diff, get_p_diff, estimate_gradient, init_gradient
     )
     return init_fn, sgld_kernel, get_params
 
@@ -276,8 +280,9 @@ def build_sgldAdam_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn(
         grad_log_post, data, batch_size
     )
+    init_diff, update_diff, get_p_diff = sgldAdam(dt, beta1, beta2, eps)
     init_fn, sgldAdam_kernel, get_params = _build_langevin_kernel(
-        *sgldAdam(dt, beta1, beta2, eps), estimate_gradient, init_gradient
+        init_diff, update_diff, get_p_diff, estimate_gradient, init_gradient
     )
     return init_fn, sgldAdam_kernel, get_params
 
@@ -307,8 +312,9 @@ def build_sgnht_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn(
         grad_log_post, data, batch_size
     )
+    init_diff, update_diff, get_p_diff = sgnht(dt, a)
     init_fn, sgnht_kernel, get_params = _build_langevin_kernel(
-        *sgnht(dt, a), estimate_gradient, init_gradient
+        init_diff, update_diff, get_p_diff, estimate_gradient, init_gradient
     )
     return init_fn, sgnht_kernel, get_params
 
@@ -340,8 +346,9 @@ def build_sgnhtCV_kernel(
     estimate_gradient, init_gradient = build_gradient_estimation_fn_CV(
         grad_log_post, data, batch_size, centering_value
     )
[…]
