+++ sgmcmcjax/diffusions.py
@@ -36,7 +36,7 @@
 #     A pytree representing the parameters extracted from `diffusion_state`, such that
 #     the invariant `params == get_params(init_fun(params))` holds true.
 # """
-from typing import Callable, Tuple
+from typing import Callable, Tuple, Union
 
 import jax.numpy as jnp
 from jax import lax, random
@@ -45,7 +45,7 @@ from .diffusion_util import diffusion, diffusion_palindrome, diffusion_sghmc
 
 
 @diffusion
-def sgld(dt: float) -> Tuple[Callable, Callable, Callable]:
+def sgld(dt) -> Tuple[Callable, Callable, Callable]:
     """SGLD diffusion
     https://www.ics.uci.edu/~welling/publications/papers/stoclangevin_v6.pdf
 
@@ -75,7 +75,7 @@ def sgld(dt: float) -> Tuple[Callable, Callable, Callable]:
 
 @diffusion
 def psgld(
-    dt: float, alpha: float = 0.99, eps: float = 1e-5
+    dt, alpha: float = 0.99, eps: float = 1e-5
 ) -> Tuple[Callable, Callable, Callable]:
     """Preconditioned SGLD diffusion
     See algorithm 1 in paper: https://arxiv.org/pdf/1512.07666.pdf
@@ -114,7 +114,7 @@ def psgld(
 
 @diffusion
 def sgldAdam(
-    dt: float, beta1: float = 0.9, beta2: float = 0.999, eps: float = 1e-8
+    dt, beta1: float = 0.9, beta2: float = 0.999, eps: float = 1e-8
 ) -> Tuple[Callable, Callable, Callable]:
     """'ADAM'-like SGMCMC diffusion. See appendix in paper: https://arxiv.org/abs/2105.13059v1
 
@@ -159,7 +159,7 @@ def sgldAdam(
 
 @diffusion_sghmc
 def sghmc(
-    dt: float, alpha: float = 0.01, beta: float = 0
+    dt, alpha: float = 0.01, beta: float = 0
 ) -> Tuple[Callable, Callable, Callable, Callable]:
     """diffusion for stochastic gradient HMC.
     See paper: https://arxiv.org/abs/1402.4102. Uses the parametrisation in section G (appendix)
@@ -203,8 +203,8 @@ def sghmc(
 
 @diffusion_palindrome
 def baoab(
-    dt: float, gamma: float, tau: float = 1.0
-) -> Tuple[Callable, Callable, Callable, Callable]:
+    dt, gamma: float, tau: float = 1.0
+) -> Tuple[Callable, Tuple[Callable, Callable], Callable]:
     """BAOAB splitting scheme for the underdampled Langevin diffusion. https://aip.scitation.org/doi/abs/10.1063/1.4802990
 
     Args:
@@ -213,7 +213,7 @@ def baoab(
         tau (float, optional): temperature. Defaults to 1.
 
     Returns:
-        Tuple[Callable, Callable, Callable, Callable]: An (init_fun, (update1, update2), get_params) triple.
+        Tuple[Callable, Tuple[Callable, Callable], Callable]: An (init_fun, (update1, update2), get_params) triple.
     """
     dt = make_schedule(dt)
 
@@ -228,7 +228,7 @@ def baoab(
         x = x + v * dt(i) * 0.5
 
         c1 = jnp.exp(-gamma * dt(i))
-        c2 = jnp.sqrt(1 - c1 ** 2)
+        c2 = jnp.sqrt(1 - c1**2)
         v = c1 * v + tau * c2 * random.normal(k, shape=jnp.shape(v))
 
         x = x + v * dt(i) * 0.5
@@ -248,7 +248,7 @@ def baoab(
 
 
[…]
