+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.10
+current_version = 0.2.11
 
 [aliases]
 test = pytest
