### 2022-02-27 jeremiecoullon/SGMCMCJax#29 @jeremiecoullon
Thanks for this suggestion!

I'll add a CONTRIBUTION.md guide following [PyMC's one](https://github.com/pymc-devs/pymc/blob/main/CONTRIBUTING.md).

ArviZ and PyMC tools: what tools do do mean?
