### 2022-02-28 jeremiecoullon/SGMCMCJax#27 @jeremiecoullon
I made these changes to paper; let me know if you think I should clarify this some more!
