### 2022-03-01 jeremiecoullon/SGMCMCJax#29 @canyon289
> ArviZ and PyMC tools: what tools do do mean?
Community engagement tools, like github discussion, gitter, discourse etc if people have questions or to reach out
