### 2022-03-01 jeremiecoullon/SGMCMCJax#29 @canyon289
Looking over everything added this is a great start https://github.com/jeremiecoullon/SGMCMCJax/blob/master/CONTRIBUTING.md

Copying from the JOSS guidelines, I cant see to find Item 3 "seek support". If I had was seeking support how would I obtain it?
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
