### 2022-03-01 #4113 @canyon289
Rereviewed after many great changes, thanks @jeremiecoullon

The only two questions I have are regarding these two JOSS requirements, API documentation and a channel for support

https://github.com/jeremiecoullon/SGMCMCJax/issues/38
https://github.com/jeremiecoullon/SGMCMCJax/issues/29
