### 2022-03-01 #4113 @jeremiecoullon
@canyon289 Thanks!
- channel for support: I forgot to add this; I'll do it this week
- API documentation: I'm still battling with readthedocs :) I'll try to fix it this week
