### 2022-03-01 3b027d6
add optimizers tests and ran black/isort
