### 2022-03-01 2574d05
add optimizers tests and ran black/isort (#52)
* add optimizers tests
* fix isort/black compatibilty
