### 2022-03-01 jeremiecoullon/SGMCMCJax#38 @jeremiecoullon
Finally fixed!
