### 2022-03-01 jeremiecoullon/SGMCMCJax#30 @jeremiecoullon
Done :)
