+++ README.md
@@ -40,6 +40,11 @@ Nsamples = 10_000
 samples = my_sampler(key, Nsamples, jnp.zeros(D))
 ```
 
+# Ask a question or open an issue
+
+Please open issues on [Github Issue Tracker](https://github.com/jeremiecoullon/SGMCMCJax/issues), or ask a question in the [Discussion section](https://github.com/jeremiecoullon/SGMCMCJax/discussions) on Github. 
+
+
 ## Samplers
 
 The library includes several SGMCMC algorithms with their pros and cons briefly discussed in the [documentation](https://sgmcmcjax.readthedocs.io/en/latest/all_samplers.html).
