### 2022-03-01 jeremiecoullon/SGMCMCJax#29 @jeremiecoullon
I just added info about support in the readme and contributing.md. Let me know if there's anything else I can fix up.
