### 2022-03-01 a3d20f8
docs: link pymc example
