### 2022-03-04 jeremiecoullon/SGMCMCJax#38 @ColCarroll
Looks great! https://github.com/openjournals/joss-reviews/issues/4113
