### 2022-03-04 jeremiecoullon/SGMCMCJax#33 @ColCarroll
https://github.com/openjournals/joss-reviews/issues/4113
