### 2022-03-04 jeremiecoullon/SGMCMCJax#28 @jeremiecoullon
I'm trying to reproduce this: I set my firefox browser to dark theme but the docs page still have a white background.

What specifically was set to black theme in your setup?
