### 2022-03-05 jeremiecoullon/SGMCMCJax#28 @canyon289
hm, I'm not sure. i tried on my laptop and its fine now.
