### 2022-03-05 jeremiecoullon/SGMCMCJax#28 @jeremiecoullon
Other libraries (for example PyMC) seem to have similar looking plots, so I'm thinking of leaving these as they are.
What do you think?
