### 2022-03-05 jeremiecoullon/SGMCMCJax#28 @canyon289
Sounds good!
