### 2022-03-06 #4113 @jeremiecoullon
@canyon289 I have closed the last of the issues; do you have any other suggested changes for the library/paper?
