### 2022-03-06 #4113 @jeremiecoullon
I made a small change in the paper: I specify that jax is the main requirement.
