+++ paper.bib
@@ -8,7 +8,7 @@
 }
 
 @article{nemeth2021stochastic,
-  title={Stochastic gradient markov chain monte carlo},
+  title={{Stochastic gradient Markov chain Monte Carlo}},
   author={Nemeth, Christopher and Fearnhead, Paul},
   journal={Journal of the American Statistical Association},
   volume={116},
@@ -41,7 +41,7 @@
 }
 
 @article{baker2019sgmcmc,
-title = {sgmcmc: An R Package for Stochastic Gradient Markov Chain Monte Carlo},
+title = {{sgmcmc}: An {R} Package for Stochastic Gradient {M}arkov chain {M}onte {C}arlo},
 author = {Jack Baker and Paul Fearnhead and Emily B. Fox and Christopher Nemeth},
 year = {2019},
 month = {oct},
