+++ paper.bib
@@ -124,13 +124,6 @@ author={
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
-@software{jax2018github,
-  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
-  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
-  url = {http://github.com/google/jax},
-  version = {0.2.5},
-  year = {2018},
-}
 
 @unpublished{murphy2023 ,
       title={Probabilistic Machine Learning: Advanced Topics},
