+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.11
+current_version = 0.2.12
 
 [aliases]
 test = pytest
