### 2022-04-14 #4113 @jeremiecoullon
@dfm : I finally finished this :)

1. I read through the paper a final time; looks good
2. Release version: 0.2.12
3. Zenodo DOI: 10.5281/zenodo.6460681

Zenodo archived link: https://zenodo.org/record/6460681

Let me know if something is missing!
